pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[[tool.uv.index]]
url = "https://mirrors.pku.edu.cn/pypi/web/simple"
default = true

[project]
name = "nodebb-python-client"
version = "1.0.0"
description = "A modern Python client for NodeBB Forum API with OpenAI Function Calling support"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "NodeBB Community"}
]
keywords = [
    "nodebb",
    "forum",
    "api",
    "client",
    "openai",
    "function-calling",
    "rest-api",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/Forum",
]

dependencies = [
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/NodeBB/NodeBB"
Documentation = "https://github.com/NodeBB/NodeBB#readme"
Repository = "https://github.com/NodeBB/NodeBB"
Issues = "https://github.com/NodeBB/NodeBB/issues"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ["py310"]
include = "\\.pyi?$"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
strict = true

[[tool.mypy.overrides]]
module = "httpx._transports.default"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --cov=nodebb --cov-report=term-missing"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    "*/__init__.py",
]

[tool.ruff]
target-version = "py310"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Imported but unused
